+++ bin/genbank-grep.py
@@ -5,19 +5,17 @@ import re
 import argparse
 from Bio import SeqIO
 
-from dark.reads import DNARead
-
 parser = argparse.ArgumentParser(
     formatter_class=argparse.ArgumentDefaultsHelpFormatter,
-    description=('Read a GenBank flat file and print FASTA for all '
-                 '(or specific) records.'))
+    description=('Read a GenBank flat file and print GenBank or FASTA for '
+                 'matching records.'))
 
 parser.add_argument(
     'genbankFile', metavar='FILE.gb', nargs='+',
     help='The GenBank file(s) to examine.')
 
 parser.add_argument(
-    '--idRegex', metavar='REGEX',
+    '--idRegex', metavar='REGEX', default='.',
     help=('Only print records whose id (or description, if --matchDescription '
           'is used) matches this regular expression.'))
 
@@ -31,6 +29,10 @@ parser.add_argument(
     help='Match on the record description as well as the id.')
 
 parser.add_argument(
+    '--format', choices=('fasta', 'gb'), default='gb',
+    help='The output format.')
+
+parser.add_argument(
     '--expectedCount', type=int,
     help=('The expected number of records to find. If this number is not '
           'found, print a message and exit non-zero.'))
@@ -42,19 +44,25 @@ parser.add_argument(
 args = parser.parse_args()
 
 recordCount = matchCount = 0
-match = getattr(re.compile(args.idRegex), args.match) if args.idRegex else None
+match = getattr(re.compile(args.idRegex), args.match)
 
 matchDescription = args.matchDescription
 
+
+def filterFunc(records):
+    global matchCount, recordCount
+    for record in records:
+        recordCount += 1
+        id_ = record.id + (
+            f" {record.description}" if record.description else "")
+        if match(id_ if matchDescription else record.id):
+            matchCount += 1
+            yield record
+
+
 for filename in args.genbankFile:
     with open(filename) as fp:
-        for record in SeqIO.parse(fp, 'gb'):
-            recordCount += 1
-            id_ = record.id + (
-                f" {record.description}" if record.description else "")
-            if match is None or match(id_ if matchDescription else record.id):
-                matchCount += 1
-                print(DNARead(id_, str(record.seq)).toString(), end='')
+        SeqIO.write(filterFunc(SeqIO.parse(fp, 'gb')), sys.stdout, args.format)
 
 if args.verbose:
     print(f'Read {recordCount} record{"s" if recordCount > 1 else ""}, '